requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
rawpy>=0.18.0
Pillow>=9.0.0
pycocotools>=2.0.6
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.10.0
scipy>=1.9.0
lxml>=4.9.0
